Move test files alongside source files and simplify vitest include pattern
