+++ vite.config.ts
@@ -22,6 +22,6 @@ export default defineConfig({
   },
   test: {
     environment: "node",
-    include: ["src/**/__tests__/**/*.test.ts"],
+    include: ["src/**/*.test.ts"],
   },
 });
